refactor(cloudflare): isolate R2 storage backends
Keep r2Storage as the single dispatch entrypoint while moving S3-backed R2 access and deprecated Wrangler-backed access into separate implementation files.

Constraint: Existing cloudflareApiToken-based r2Storage configuration must remain compatible but be marked deprecated.

Rejected: Leaving both storage backends in r2Storage.ts | it made future Wrangler removal harder to review.

Confidence: high

Scope-risk: narrow

Directive: Remove the Wrangler backend by deleting r2WranglerStorage.ts and the r2Storage fallback branch when legacy support ends.

Tested: pnpm exec oxfmt plugins/cloudflare/src/r2Storage.ts plugins/cloudflare/src/r2S3Storage.ts plugins/cloudflare/src/r2WranglerStorage.ts; pnpm --filter @hot-updater/cloudflare test:type; pnpm --filter @hot-updater/cloudflare build; pnpm -w test -- --project='unit:@hot-updater/cloudflare'; .agents/skills/fix-ci/scripts/run_fixci.sh all

Not-tested: Live Cloudflare R2 upload/download against a real account.
